# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CRUD - Phase 2 | Edacy",
        "description": "Documentation de L'API d'un CRUD pour validation de la phase 2 du programme propose par Edacy",
        "contact": {
            "name": "Issakha CISSE",
            "email": "admin@edacy.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Adresse de l'API"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Liste des produits",
                "description": "Récupère la liste de tous les produits",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Liste des produits récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Créer un produit",
                "description": "Crée un nouveau produit",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "quantity"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom du produit",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "iPhone 15"
                                    },
                                    "description": {
                                        "description": "Description du produit",
                                        "type": "string",
                                        "example": "Dernier smartphone Apple",
                                        "nullable": true
                                    },
                                    "price": {
                                        "description": "Prix du produit",
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 999.99
                                    },
                                    "quantity": {
                                        "description": "Quantité en stock",
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Produit crée avec succés !"
                                        },
                                        "product": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Données érronées."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Ce champ est obligatoire"
                                                    }
                                                },
                                                "price": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Ce champ est obligatoire"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Afficher un produit",
                "description": "Récupère les détails d'un produit spécifique",
                "operationId": "getProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du produit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Modifier un produit",
                "description": "Met à jour les informations d'un produit existant",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nom du produit",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "iPhone 15 Pro"
                                    },
                                    "description": {
                                        "description": "Description du produit",
                                        "type": "string",
                                        "example": "Dernier smartphone Apple Pro",
                                        "nullable": true
                                    },
                                    "price": {
                                        "description": "Prix du produit",
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 1199.99
                                    },
                                    "quantity": {
                                        "description": "Quantité en stock",
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Produit modifié avec succés"
                                        },
                                        "product": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Supprimer un produit",
                "description": "Supprime un produit de la base de données",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produit supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "oneOf": [
                                            {
                                                "type": "string",
                                                "example": "Produit supprimé avec succés"
                                            },
                                            {
                                                "type": "null"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Modèle produit",
                "required": [
                    "id",
                    "name",
                    "price",
                    "quantity",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "ID du produit",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nom du produit",
                        "type": "string",
                        "example": "iPhone 15"
                    },
                    "description": {
                        "description": "Description du produit",
                        "type": "string",
                        "example": "Dernier smartphone Apple",
                        "nullable": true
                    },
                    "price": {
                        "description": "Prix du produit",
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "quantity": {
                        "description": "Quantité en stock",
                        "type": "integer",
                        "example": 10
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilisez le token retourné lors du login. Format: Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
...
}